Laravel advance 2

+++ database/migrations/2023_02_06_074441_edit_type_role_users_table.php
@@ -25,8 +25,10 @@ class EditTypeRoleUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
-        });
+        if (Schema::hasColumn('users', 'role')){
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('role');
+            });
+        }
     }
 }
